requirements.txt
scrapy-playwright==0.0.43